# backend/ent/schema/Employee.go
package schema

import (
 "github.com/facebookincubator/ent"
 "github.com/facebookincubator/ent/schema/edge"
 "github.com/facebookincubator/ent/schema/field"
)

// Employee holds the schema definition for the Employee entity.
type Employee struct {
 ent.Schema
}

// Fields of the Employee.
func (Employee) Fields() []ent.Field {
	return []ent.Field{
		field.String("Email"),
		field.String("Name"),
		field.String("User_id"),
	}
}

// Edges of the Employee.
func (Employee) Edges() []ent.Edge {
 return []ent.Edge{
 edge.To("area", Area.Type),
 }
}